refactor: better approach to compiling Swoole extension
Per https://www.swoole.co.uk/article/docker.html, this patch modifies
each dockerfile to download a tarball and use the standard
phpize/configure/make/make install approach to building the extension.
This allows better visibility into errors when building images.

Additionally, this patch recognizes that curl is already available, and
being used to retrieve the composer signature. Interestingly, we were
then fetching the composer installer using... PHP. This updates that
code to use curl instead.

The Swoole version to use is passed in using a build argument. The
Makefile defines a default of 4.2.12 currently, but the value may be
passed when creating images.